update trigger package
