+++ src/InfiniteStairs.cpp
@@ -14,7 +14,7 @@
 //
 
 #include "plugin.hpp"
-#include "dsp/triggers.hpp"
+#include "dsp/trigger.hpp"
 #include "dsp/ricoh_2a03.hpp"
 #include "engine/chip_module.hpp"
