+++ src/Jairasullator.cpp
@@ -14,7 +14,7 @@
 //
 
 #include "plugin.hpp"
-#include "dsp/triggers.hpp"
+#include "dsp/trigger.hpp"
 #include "dsp/general_instrument_ay_3_8910.hpp"
 #include "engine/chip_module.hpp"
