+++ src/MiniBoss.cpp
@@ -15,7 +15,7 @@
 
 #include <functional>
 #include "plugin.hpp"
-#include "dsp/triggers.hpp"
+#include "dsp/trigger.hpp"
 #include "dsp/yamaha_ym2612/feedback_operator.hpp"
 #include "engine/yamaha_ym2612_params.hpp"
