+++ src/SuperADSR.cpp
@@ -15,7 +15,7 @@
 
 #include <limits>
 #include "plugin.hpp"
-#include "dsp/triggers.hpp"
+#include "dsp/trigger.hpp"
 #include "dsp/sony_s_dsp/adsr.hpp"
 
 // ---------------------------------------------------------------------------
