+++ src/dsp/trigger.hpp
@@ -1,0 +1,25 @@
+// Triggers for detecting events in time-domain signals.
+// Copyright 2020 Christian Kauten
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+//
+
+#ifndef DSP_TRIGGER_HPP
+#define DSP_TRIGGER_HPP
+
+#include "trigger/boolean.hpp"
+#include "trigger/divider.hpp"
+#include "trigger/hold.hpp"
+#include "trigger/threshold.hpp"
+#include "trigger/zero.hpp"
+
+#endif  // DSP_TRIGGER_HPP
